+++ apps/api/src/app.module.ts
@@ -4,8 +4,16 @@ import { HealthModule } from "./health/health.module"
 import { NotificationsModule } from "./notifications/notifications.module"
 import { OtpModule } from "./otp/otp.module"
 import { ProvidersModule } from "./providers/providers.module"
+import { TemplatesModule } from "./templates/templates.module"
 
 @Module({
-    imports: [AuthModule, HealthModule, NotificationsModule, OtpModule, ProvidersModule],
+    imports: [
+        AuthModule,
+        HealthModule,
+        NotificationsModule,
+        OtpModule,
+        ProvidersModule,
+        TemplatesModule,
+    ],
 })
 export class AppModule {}